refactor(KUI-1024): add parameter fromTerm to getDetailedInfo
