+++ server/controllers/memoCtrlHelpers.js
@@ -2,6 +2,7 @@
 
 const log = require('@kth/log')
 const { getCourseRoundTerms } = require('../koppsApi')
+const { getCurrentTerm } = require('../utils/term')
 
 function enrichMemoDatasWithOutdatedFlag(memoDatas = [], roundInfos = []) {
   if (!Array.isArray(memoDatas)) {
@@ -100,17 +101,6 @@ function getIsMemoOutdated(offerings, startSelectionYear, memoData) {
 
   // Course memo does not meet the criteria
   return true
-}
-
-function getCurrentTerm(overrideDate) {
-  const JULY = 6
-  const SPRING = 1
-  const FALL = 2
-  const currentDate = overrideDate || new Date()
-  const currentYear = currentDate.getFullYear()
-  const currentMonth = currentDate.getMonth()
-  const currentSemester = currentMonth < JULY ? SPRING : FALL
-  return `${currentYear * 10 + currentSemester}`
 }
 
 async function findStartDateForMemo(memo, roundInfos) {
